Update resource state unit tests
